Revert automatic catching of exception in #stopOnError and #error.
The auto error handling was too aggressive. In certain cases, it caught
and suppressed legitimate exceptions (like the one thrown by #toArray on
errors) that was not actually caused by the argument function. Since the
user already has a way to signal errors in his error handler (via the
second parameter passed to that handler), the simplest way to fix this
is to stop catch user exceptions in that case.