Adding a Stream.splitByWithTimeout
The current splitBy method will leave the buffer unsent indefinetly if the separator is not detected.
For quick moving streams, this doesn't cause a problem but for some cases, with slower streams,
we want the buffer contents to be sent after a specific amount of time. Otherwise we are eternally one
step behind.